+++ src/plugins/extended/image.ts
@@ -1,92 +1,30 @@
-import type { Plugin, PDFRenderProps, UIRenderProps, PropPanelWidgetProps, Schema } from '@pdfme/common';
+import type { Plugin } from '@pdfme/common';
 import { image as component } from '@pdfme/schemas';
-
-interface ExtendedImage extends Schema {
-  field_key?: string;
-  field_format?: string;
-  field_options?: string;
-  readOnly?: boolean;
-}
-
-const pdfRender = async (arg: PDFRenderProps<ExtendedImage>) => {
-  const { schema } = arg;
-  if (schema.type === 'extendedImage') {
-    // Handle custom property if needed
-  }
-  await component.pdf(arg as PDFRenderProps<Schema>);
-};
-
-const uiRender = async (arg: UIRenderProps<ExtendedImage>) => {
-  const { schema } = arg;
-  if (schema.type === 'extendedImage') {
-    // Handle custom property if needed
-  }
-  await component.ui(arg as UIRenderProps<Schema>);
-};
+import { createPDFRender, createUIRender, createSchemaFunction, ExtendedSchema } from './schemaUtils';
 
 const createExtendedImageSchema = (
   fieldKeyOptions: { label: string; options: { label: string; value: string }[] }[],
   fieldFormatOptions: { label: string; value: string }[]
 ) => {
-  const defaultSchema: Schema = {
+  const defaultSchema: ExtendedSchema = {
     ...component.propPanel.defaultSchema,
     type: 'extendedImage',
   };
 
-  const schema: PropPanelWidgetProps = (props: Omit<PropPanelWidgetProps, 'rootElement'>) => {
-    const { options, activeSchema, i18n } = props;
-
-    return {
-      readOnly: {
-        title: 'Read Only',
-        type: 'boolean',
-        widget: 'switch',
-        span: 8,
-      },
-      divider1: {
-        widget: 'divider',
-        span: 24,
-      },
-      field_key: {
-        title: 'Field Key',
-        type: 'string',
-        widget: 'select',
-        span: 16,
-        props: {
-          options: fieldKeyOptions,
-          allowClear: true,
-        },
-      },
-      field_format: {
-        title: 'Format',
-        type: 'string',
-        widget: 'select',
-        span: 8,
-        props: {
-          options: fieldFormatOptions,
-          allowClear: true,
-        },
-      },
-      field_options: {
-        title: 'Format Options',
-        type: 'string',
-        widget: 'input',
-        span: 16,
-      },
-      divider2: {
-        widget: 'divider',
-        span: 24,
-      },
-      ...component.propPanel.schema,
-    };
-  };
+  const schemaFunction = createSchemaFunction(component.propPanel.schema, fieldKeyOptions, fieldFormatOptions);
 
   const propPanel = {
     defaultSchema,
-    schema
+    schema: schemaFunction,
+    widgets: component.propPanel.widgets,
   };
 
-  const extendedSchema: Plugin<ExtendedImage> = { pdf: pdfRender, ui: uiRender, propPanel };
+  const extendedSchema: Plugin<ExtendedSchema> = {
+    pdf: createPDFRender(component),
+    ui: createUIRender(component),
+    propPanel,
+  };
+
   return extendedSchema;
 };
